Changed reported SolarBoard voltages to be `f32`.

+++ electrical/system.h
@@ -896,7 +896,7 @@ pack_push
     struct MainFlightComputerDebugPacket
     {
         u32                                            timestamp_us;
-        i16                                            solarboard_voltages[2];
+        f32                                            solarboard_voltages[2];
         typeof(enum MainFlightComputerDebugStatusFlag) flags;
         u8                                             crc;
     };
